% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{iccPlot}
\alias{iccPlot}
\title{Plot expected and observed table from SitemFit}
\usage{
iccPlot(grp, itemName, ..., width = 3, dataBins = 11, basis = c(1),
  factor = 1)
}
\arguments{
\item{grp}{an IFA group}

\item{itemName}{name of item to plot}

\item{...}{Not used.  Forces remaining arguments to be specified by name.}

\item{width}{sets the x axis to [-width,width]}

\item{dataBins}{number of partitions for the latent scores}

\item{basis}{the basis vector in the latent space}

\item{factor}{the score to use (TODO: should be a function of the basis vector?)}
}
\description{
WARNING: This function is under development. The API may change in a future release.
}

