% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{itemModelExplorer}
\alias{itemModelExplorer}
\title{A Shiny app to experiment with item models}
\usage{
itemModelExplorer()
}
\description{
A Shiny app to experiment with item models
}
\examples{
\dontrun{
itemModelExplorer()  # will launch a browser in RStudio
}
}
