% Generated by roxygen2 (4.0.2): do not edit by hand
\name{itemResponseMap}
\alias{itemResponseMap}
\title{Create item response map table}
\usage{
itemResponseMap(grp, ..., factor = 1)
}
\arguments{
\item{grp}{an IFA group}

\item{...}{Not used.  Forces remaining arguments to be specified by name.}

\item{factor}{which factor to plot (defaults to 1)}
}
\description{
Categories are placed at the mean score of the examinees who picked
that category.
}

