% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apps.R
\name{modelBuilder}
\alias{modelBuilder}
\title{A Shiny app for building IFA models}
\usage{
modelBuilder()
}
\description{
A Shiny app for building IFA models
}
\examples{
\dontrun{
modelBuilder()  # will launch a browser in RStudio
}
}

