% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotInformation}
\alias{plotInformation}
\title{Plot item information in the latent distribution}
\usage{
plotInformation(grp, ..., width = 3, showTotal = FALSE, basis = c(1))
}
\arguments{
\item{grp}{an IFA group}

\item{...}{Not used.  Forces remaining arguments to be specified by name.}

\item{width}{the plot will span from -width to width}

\item{showTotal}{whether to plot the total item information}

\item{basis}{the basis vector (for multidimensional items)}
}
\description{
For multidimensional items, you will need to supply a basis
vector. This vector is normalized to unit length.
}

