% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{replicateModelBy}
\alias{replicateModelBy}
\title{Replicate a model for each group of data}
\usage{
replicateModelBy(tmpl, fullData, mMat, covMat, ..., splitCol = "population",
  refGroup = "general", split = TRUE, compressData = TRUE)
}
\arguments{
\item{tmpl}{an OpenMx model}

\item{fullData}{the complete data including the column indicating group membership}

\item{mMat}{an MxMatrix for latent means}

\item{covMat}{an MxMatrix for latent covariance}

\item{...}{Not used.  Forces remaining arguments to be specified by name.}

\item{splitCol}{the name of the column used to indicate group membership}

\item{refGroup}{the name of the reference group}

\item{split}{whether to split the data (defaults to TRUE)}

\item{compressData}{whether to apply compressDataFrame (defaults to TRUE)}
}
\description{
The reference group is fixed to a zero mean and identity covariance matrix.
}

