% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addExploratoryFactors}
\alias{addExploratoryFactors}
\title{Adds exploratory factors to a single factor model}
\usage{
addExploratoryFactors(model, toAdd, ..., addUniquenessPrior = TRUE)
}
\arguments{
\item{model}{a single factor (possibly multigroup) model}

\item{toAdd}{the number of factors to add}

\item{...}{Not used.  Forces remaining arguments to be specified by name.}

\item{addUniquenessPrior}{whether to add a uniqueness prior to the model (default TRUE)}
}
\description{
Adds exploratory factors to a single factor model
}

