% Generated by roxygen2 (4.0.2): do not edit by hand
\name{itemModelExplorer}
\alias{itemModelExplorer}
\title{A Shiny app to experiment with item models}
\usage{
itemModelExplorer()
}
\description{
A Shiny app to experiment with item models
}
\examples{
\dontrun{
itemModelExplorer()  # will launch a browser in RStudio
}
}

