% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{itemResponseMap}
\alias{itemResponseMap}
\title{Create item response map table}
\usage{
itemResponseMap(grp, ..., factor = 1)
}
\arguments{
\item{grp}{an IFA group}

\item{...}{Not used.  Forces remaining arguments to be specified by name.}

\item{factor}{which factor to plot (defaults to 1)}
}
\value{
A data.frame of the raw data backing the plot. Item outcomes
without any observations are omitted.
}
\description{
Categories are placed at the mean score of the examinees who picked
that category.
}
