% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{SitemPlot}
\alias{SitemPlot}
\title{Plot expected and observed table from SitemFit}
\usage{
SitemPlot(sout, itemName, ..., showSampleSize = TRUE)
}
\arguments{
\item{sout}{output from SitemFit}

\item{itemName}{name of item to plot}

\item{...}{Not used.  Forces remaining arguments to be specified by name.}

\item{showSampleSize}{whether to show the sample size at the top of the plot}
}
\description{
Plot expected and observed table from SitemFit
}

